% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_definition.R
\docType{data}
\name{ad_metrics}
\alias{ad_metrics}
\title{Advertisng world metrics}
\format{An object of class \code{quosures} of length 11.}
\usage{
ad_metrics
}
\description{
Advertisng world metrics
}
\keyword{datasets}
