% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmetrics.R
\name{add}
\alias{add}
\title{Add metrics to data.frame}
\usage{
add(df, ..., metrics = ad_metrics, summarize = FALSE)
}
\arguments{
\item{df}{data.frame}

\item{...}{group keys}

\item{metrics}{metrics}

\item{summarize}{summarize all data or not (mutate compatible behavior) when group keys(thee dots) are empty}
}
\description{
Add metrics to data.frame
}
