% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmetrics.R
\name{add_metrics}
\alias{add_metrics}
\title{Add metrics to data.frame}
\usage{
add_metrics(df, ..., metrics = default_metrics)
}
\arguments{
\item{df}{data.frame}

\item{...}{dots}

\item{metrics}{metrics}
}
\description{
Add metrics to data.frame
}
