% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{binary_acc}
\alias{binary_acc}
\title{binary_acc}
\usage{
binary_acc(preds, target, threshold = 0.5, multidim_average = "global")
}
\arguments{
\item{preds}{Predicted label or predicted probability between 0 and 1, same shape as target label}

\item{target}{Target label}

\item{threshold}{The numerical cut-off between 0 and 1 to transform predicted probability into binary predicted labels}

\item{multidim_average}{Average model: global-average across all accuracies, samplewise-average across the all but the first dimensions (calculated independently for each sample)}
}
\value{
Binary accuracy for preds and target, with format dictated by multidim_average command.
}
\description{
Calculate the binary class accuracy
}
