% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_f1.R
\name{binary_f1}
\alias{binary_f1}
\title{binary_f1}
\usage{
binary_f1(preds, target, threshold = 0.5, multidim_average = "global")
}
\arguments{
\item{preds}{Predicted labels or predicted probability between 0 and 1,
same shape as target label}

\item{target}{Target label}

\item{threshold}{The numerical cut-off between 0 and 1 to transform
predicted probability into binary predicted labels}

\item{multidim_average}{Average model: global-average across all f1s,
samplewise-average across the all but the first dimensions (calculated
independently for each sample)}
}
\value{
f1score value for binary preds and target input, with format dictated by
multidim_average command.
}
\description{
Calculate the binary classification f1-score for a given predicted set of
values and corresponding targets. In other words, this function estimate how accurate
the true prediction value by the model is.
}
\examples{
binary_f1(c(0.8, 0.2), c(1,1), 0.3)
binary_f1(c(1,1), c(0,1))
}
