% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_precision.R
\name{binary_precision}
\alias{binary_precision}
\title{binary_precision}
\usage{
binary_precision(preds, target, threshold = 0.5, multidim_average = "global")
}
\arguments{
\item{preds}{Predicted labels or predicted probability between 0 and 1,
same shape as target label}

\item{target}{Target label}

\item{threshold}{The numerical cut-off between 0 and 1 to transform
predicted probability into binary predicted labels}

\item{multidim_average}{Average model: global-average across all accuracies,
samplewise-average across the all but the first dimensions (calculated
independently for each sample)}
}
\value{
Binary precision value for preds and target, with format dictated by
multidim_average command.
}
\description{
Calculate the binary classtypeification precision for a given predicted set of
values and corresponding targets. In other words, this function estimate how accurate
the true prediction value by the model is.
}
\examples{
binary_precision(c(0.8, 0.2), c(1,1), 0.3)
binary_precision(c(1,1), c(0,1))
}
