% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{confusion_scores}
\alias{confusion_scores}
\title{confusion_scores}
\usage{
confusion_scores(preds, target, multidim_average = "global")
}
\arguments{
\item{preds}{Predicted label, same shape as target label}

\item{target}{Target label}

\item{multidim_average}{Average model: global-average across all accuracies, samplewise-average across the all but the first dimensions (calculated independently for each sample)}
}
\value{
A list contains matrix-confusion matrix, tp-true positives, fn-false negatives, fp-false positives, tn-true negatives
}
\description{
Calculate confusion matrix
}
