% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmetrics.R
\name{define}
\alias{define}
\title{Define metrics}
\usage{
define(...)
}
\arguments{
\item{...}{Metrics definition.

These arguments are automatically \link[rlang:quo]{quoted}
and \link[rlang:eval_tidy]{evaluated} in the context of the data frame.}
}
\description{
This helper is just synonum for \link[rlang:quos]{rlang::quos} intended to provide seamless experience for package user.
}
\seealso{
\link[rlang:quos]{quos}, \href{https://cran.r-project.org/package=dplyr/vignettes/programming.html}{dplyr's vignettes}
}
