% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmetrics.R
\name{disaggregate}
\alias{disaggregate}
\title{Disaggregate metrics defined as aggregate function}
\usage{
disaggregate(metrics)
}
\arguments{
\item{metrics}{metrics defined by mmetrics::define()}
}
\value{
disaggregated metrics (rlang::quosure or rlang::quosures)
}
\description{
Disaggregate metrics defined as aggregate function
}
\examples{

metrics <- mmetrics::define(
  cost = sum(cost),
  ctr  = sum(click)/sum(impression)
)

mmetrics::disaggregate(metrics)

}
