% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dummy_data}
\alias{dummy_data}
\title{Dummy data.frame for this package}
\format{A data frame with 10 rows and 6 variables:
\describe{
\item{gender}{gender, Men (M) or Female (F)}
\item{age}{age}
\item{cost}{the amount how much money do you earn or advertisers spend}
\item{impression}{the number of how many times some ads are shown to users}
\item{click}{the number of how many times users click ads}
\item{conversion}{the number of how many times users converge}
}}
\usage{
dummy_data
}
\description{
A dataset contains columns often used in the digital marketing industry by user demographics
}
\keyword{datasets}
