% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmetrics.R
\name{mfilter}
\alias{mfilter}
\title{Pick evaluable metrics in the given data frame}
\usage{
mfilter(metrics, df)
}
\arguments{
\item{metrics}{Metrics}

\item{df}{Data frame}
}
\value{
Evaluable metrics
}
\description{
Pick evaluable metrics in the given data frame
}
