% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot.R
\name{mplot_bar}
\alias{mplot_bar}
\title{Plot bar charts}
\usage{
mplot_bar(df, y, x = NULL)
}
\arguments{
\item{df}{data.frame}

\item{y}{y axis}

\item{x}{x axis}
}
\value{
ggplot object
}
\description{
Plot bar charts.
}
\examples{
\dontrun{
df <- mmetrics::dummy_data
# Add metrics and plot directly
mmetrics::mplot_bar(mmetrics::add(df, gender), ctr, gender)
# You can remove x parameter. in this case first column is assumed as x parameter
mmetrics::mplot_bar(mmetrics::add(df, gender), ctr)
}
}
