% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{multiclass_acc}
\alias{multiclass_acc}
\title{multiclass_acc}
\usage{
multiclass_acc(
  preds,
  target,
  num_class,
  multidim_average = "global",
  average = "micro"
)
}
\arguments{
\item{preds}{Predicted label with the same shape as target label, or
predicted probability between 0 and 1 for each class that has one
additional dimension compared with target label}

\item{target}{Target label that has been transformed into dinstinct integers
to refer to each class}

\item{num_class}{Total number of classes to be taken into consideration
when "average"="macro"}

\item{multidim_average}{Average model: global-average across all accuracies,
samplewise-average across the all but the first dimensions (calculated
independently for each sample)}

\item{average}{Defines the reduction that is applied over labels.
Micro-sum over all class labels.
Macro-calculate class label-wise statistics and then take the average.
The parameter average makes a difference in calculation only when the accuracy
is required on a global level}
}
\value{
Multiclass accuracy for preds and target, with format dictated by
multidim_average command and average methods choice.
}
\description{
Calculate the multiclass accuracy for a given predicted set of
values and corresponding targets
}
\examples{
y_pred = matrix(c(0.1, 0.5, 0.4, 0.9, 0.2, 0.8), 2,3)
y_target = c(2,1)
multiclass_acc(y_pred, y_target, 3)
}
