% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass_f1.R
\name{multiclass_f1}
\alias{multiclass_f1}
\title{multiclass_f1}
\usage{
multiclass_f1(preds, target, multidim_average = "global", average = "micro")
}
\arguments{
\item{preds}{Predicted label with the same shape as target label, or
predicted probability between 0 and 1 for each class that has one
additional dimension compared with target label}

\item{target}{Target label}

\item{multidim_average}{Average model: global-average across all f1 scores,
samplewise-average across the all but the first dimensions (calculated
independently for each sample)}

\item{average}{Defines the reduction that is applied over labels.
Micro-sum over all class labels, that is all true positives for each class divided
by all positive predicted values for each class.
Macro-calculate class label-wise f1 scores and then take the average.}
}
\value{
Multiclass f1-score for preds and target, with format dictated by
multidim_average argument and average methods choice.
}
\description{
Calculate the multiclass f1 value for a given predicted set of
values and corresponding targets
}
\examples{
y_pred = matrix(c(0.1, 0.5, 0.4, 0.9, 0.2, 0.8), 2,3)
y_target = c(2,1)
multiclass_f1(y_pred, y_target)
}
