% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiclass_recall.R
\name{multiclass_recall}
\alias{multiclass_recall}
\title{multiclass_precision}
\usage{
multiclass_recall(
  preds,
  target,
  multidim_average = "global",
  average = "micro"
)
}
\arguments{
\item{preds}{Predicted label with the same shape as target label, or
predicted probability between 0 and 1 for each class that has one
additional dimension compared with target label}

\item{target}{Target label that has been transformed into dinstinct integers
to refer to each class}

\item{multidim_average}{Average model: global-average across all recall scores,
samplewise-average across the all but the first dimensions (calculated
independently for each sample)}

\item{average}{Defines the reduction that is applied over labels.
Micro-sum over all class labels, that is all true positives for each class divided
by all positive predicted values for each class.
Macro-calculate class label-wise recall scores and then take the average.}
}
\value{
Multiclass recall for preds and target, with format dictated by
multidim_average argument and average methods choice.
}
\description{
Calculate the multiclass recall value for a given predicted set of
values and corresponding targets
}
\examples{
y_pred = matrix(c(0.1, 0.5, 0.4, 0.9, 0.2, 0.8), 2,3)
y_target = c(2,1)
multiclass_recall(y_pred, y_target)
}
