% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmetrics.R
\name{mfilter}
\alias{mfilter}
\title{Filter metrics which is evaluatable}
\usage{
mfilter(metrics, df)
}
\arguments{
\item{metrics}{metrics}

\item{df}{data.frame}
}
\description{
Filter metrics which is evaluatable
}
